Correctly set `settingsVersion`.
This fixes an error in #1950.

There, we need to detect new installs, and did so by checking
`settingsVersion`.  However, `settingsVersion` is *always* truthy.  So
the "new install" code never triggers.  Here, the default value for
`settingsVersion` is falsy.

Note, we *always* set `settingsVersion` when the extension is loaded.
So `settingsVersion` is always set.