Fix `<count>removeTab`.
There's a nasty little bug in `removeTab` when you remove more tabs than
there are in the window (and there is a second window):

- all of the tabs in the currently-focused window are removed
- then, later (so, time passes), when you change tab in the other
  window, we begin removing tabs again!

The source of the bug is our reliance on
`chrome.tabs.onSelectionChanged`, which doesn't first when removing the
last tab in a window (and there is another window).

Regardless, the previous semantics of `<count>removeTab` was
questionable.  It was preactically impossible to predict which tabs
would removed.  This picks the current tab, then those to the right of
the current tab, then those to the left.