Add install date to logging page.
This implements a poor-man's build info.

See #1352.  Unfortunately, that requires a separate build target, and
does not work with `cake autobuild`.

This just records the *install date* and displays that info on the
logging page.  "Install date" because we can reliably determine it, and
because it does answer the question *have I upgrade Vimium on this
machine since last week?*.  And on the logging page because that's out
of the way, and not part of the regular Vimium interface.