Fix for <count>F.
We were immediately restarting link-hints mode if a count was present.
Unfortunately, that meant that we were detecting our own link-hint click
and exiting immediately.  So, with a count of 6, we were only getting 3
link-hint activations.

To avoid this, we add a short delay (just nextTick).

Also, move some other stuff arund to make sure this works in all cases
(e.g. wait-for-enter).