Refactor hints; fix exit sequence.
Previously, the exit sequence when a link was "clicked" was spread over
several functions with several callbacks.  This made it difficult to
verify that the correct actions were happening in the correct order.
Indeed, they weren't in at least one case (we were still showing hints
while "waiting for enter").

This fixes that by putting all of the various deactivation orders into
one place, `@activateLink()`, and simplifies `@deactivateMode()`
accordingle.