Refactor hints; add `userMightOverType`.
Previously, we set a variable `delay` and then did some logical
gymnastics to get the correct effect.

However, in fact, all we care about is whether the user might over-type
the links text.  So changing to using that as a Boolean flag greatly
simplifies the logic.  And we lose about 10 LoC.