Make command names in the help diablog clickable.
Clicking on command names in the help dialog copies the command name to
the clipboard.  Moreover, command names can be clicked (hence copied)
with link hints.

As a a side effect, any element on any page with the `vimiumClickable`
class becomes clickable.  That's zero elements on zero pages, currently.
But there was a PR some time ago where a web developer was looking for a
way to make their elements vimium clickable.

On the negative side, this is pretty undiscoverable.  Also, the
`indexOf()` class-name test within a string isn't exactly precise.