Generalise topFrame commands.
This generalises the mechanism by which commands are always run in the
tab's main/top frame.  Currently, that's just the Vomnibar.

This is a precursor to moving other UI components to the main/top frame.
It should be fairly trivial to move the help page to the main frame.
The HUD might be trickier.

Mention: @mrmr1993.