Help dialog to top frame; initial version.
This moves the help dialog such that it is always shown in the main/top
frame, just like the Vomnibar.

This works.  But...

There is a problem.  When a UI component closes, Chrome focuses some
other frame.  Previously, (with just the Vomnibar) we could predict that
it would be the main/top frame, and we could wait for that frame to
receive the focus before re-focusing the original frame.

Here, however, we can open the Vomnibar from within the help dialog.  In
this case, if we then close all of the frames with `Esc`, the focus ends
up in the (hidden!) help dialog.  This is not actually a new problem.
It exists in `master` too.

In this commit, there is a big "FIXME".  We get around the issue by
waiting an aribitrary short time (150ms) before sending the request to
re-focus the original frame.  So there's a pretty serious race
condition.  Also, during that waiting period, keyboard events are being
sent to some unknown frame.