Tweak options page class names...
We were using the class name exclusionRemoveButton for the `td`
containing the button to remove exclusion rules on the options page.

Because of this, we were (correctly) picking up the button itself as
clickable, but incorrectly picking up also its enclosing `td` (which
isn't actually clickable at all).  The effect was to always have two
hint labels over the remove button.

This was happening simply because we have the text "button" in the class
name.  So, here, we just rename the classes.