Add options-page button to sort exclusion rules.
After a while, the list of exclusion rules on the options page can
become quite long, and it can be difficult to find the rule (or rules)
pertaining to a particular site.

This adds a button to the options page to sort the rules.  The sort
order currently is...  strip any leading `protocol://`, then strip any
leading `www.`, then sort what's left.  So `https://github.com` and
`http://www.github.com` sort beside each other.