Relocate first scrollable element.
When the scrollers `activatedElement` isn't scrollable, we currently
search up the DOM tree for the next scrollable thing, and stop at
`document.body`.  However, it can be that that isn't scrollable.

When that happens, this commit searches for a scrollable element in the
same way as it does when the frame initially loads.  This makes it
possible to restart scrolling on pages like this one:

  - http://redux.js.org/docs/basics/UsageWithReact.html

after clicking one of index links.

Inspired by example posted by @marlun in #425.