Refactor the front-end initialisation sequence.
The front-end initialisation sequence has become quite confused.

This simplifies it, makes things which must be idemtpotent explicit and
renames some functions to make it clear when they run.  It also avoids a
situation where we were possibly installing a `domReady` function to
initialise the HUD multiple times.

Should be a no-op functionality wise.