Rework global mark activation.
There were two problems, both stemming from the fact that the
notification was being displayed in the top frame, even if the mark was
triggered in another frame:

1. That looks odd, because we close the HUD in one frame then open it in
   another.

2. As a side effect, we were moving the focus to the top frame.

Here, we work out what's going to happen before sending the message to
the background page.  This allows us to display the message in the HUD
in the frame which generated it.