Ensure correct sendResponse() is called.
With several frames, only one frame can call sendResponse(), whichever
one happens to call it first.

getScrollPosition, is the only handler for which we care about the
response.  Here, we make sure that sendResponse() will *only* be called
in the top frame - where we want it called.

This fixes a possible race condition in global marks.

Additionally, although it's not the primary intention here, this also
avoids calling sendResponse() unnecessarily in cases where we don't use
the response.