Escape closes help dialog.
With the help dialog in an iframe, Escape no longer closes it if that
iframe loses the focus.  This fixes that.

See point 2 of #2045.

This is not a perfect solution: it only works if the focus ends up in
the frame from which the help dialog was launched.  However, that is the
the common case and, in particular, it is the case which arises on the
options page -- which is a particularly important use case.