Do not hide help dialog on click (on options page).
Normally, we close the help dialog when the user clicks outside of it.

On the options page, however, the user might want the help dialog open
while they type command names into the key-mappings input.

Therefore, we disable hide-on-click for this specific use case.

This is a rather unfortunate hack.  However, #2045 is a serious problem
for an important use case.

Fixes #2045.