Change hintMarkers to @hintMarkers.
hintMarkers was previously passed around (within the link-hints mode
class) from function to function.  With #2048 (global link hints), it is
better if this becomes @hintMarkers (that is, is available directly to
all methods within the class.

With #2048, we used bind to avoid having to do this - to keep the diff
clearer.  Here, we do it directly.