Do not blur focused input on URL change.
Currently, if grab-back-focus is enabled, then we grab back the focus on
a URL change -- including bluring any already focused input.

On some pages, e.g. Google Inbox, it is that act of focusing an input
which triggers a URL change, and in such cases, we should not blur the
input.

This is a refinement of #1622.