Escape to exit WaitForEnter.
Context: filtered link hints with wait-for-enter enabled.  Once a link
is selected, it is highlighted.  We then consume all keyboard events
until the user hits enter, at which point the link is activated.

Problem:  When we're waiting, the link is highlighted.  It looks to the
user like `Escape` should cancel.

This implements escape to cancel at that point in the exit sequence.