Ensure settings are loaded.
With global link hints, hints might be launched in one frame when the
settings are not yet loaded in another.  This could lead to one frame
using one settings value, and another another value.  (Because we use
the default value if the settings are not yet loaded.)  And this in turn
could cause link hints to crash if filetered hints are used (because, in
that case, we would not calculate hint.linkText in one frame, but then
try to use that value later in another frame).