Separate registerFrame from port initialization.
This is a no-op.

It separates the process of registering a frame with the background page
from the port initialization.  The idea is that -- soon -- we will only
register some frames; in particular, we should not register the many
tiny iframes on pages like GMail.

The eventual goal is to speed up the global link-hints initialisation
sequence.