Rework DomUtils.documentReady().
There seems to be an issue on this page:

   - http://i.imgur.com/PdmUjij.jpg

whereby "DOMContentLoaded" isn't firing.  The page structure is unusual
(involving a shadow-DOM element), which might be the source of the
problem.

What happens is that the "DOMContentLoaded" event fires as required, but
the document state is still "loading".

Here, we just say that if the "DOMContentLoaded" handler has fired
once, then we're good to go.