Make document.body selectable as scrollable.
Previously, we could select divs, uls, and ols for scrolling, but we
couldn't get back to scrolling the document body.

This makes it possible to select document.body for scrolling.

Unfortunately, sometimes the hint appears in a rather odd place (because
it's "on top of" something else which is clickable.