Better handling of port failures.
If we lose contact with a frame (it's goine away) while still awaiting
hints from that frame, then post a dummy/empty list of hints instead.

This seems very unlikely to come up, but we should guard against it
anyway.

We use "nextTick()" so that we finish sending the current message
before sending the dummy hints.