Add backstop for global link hints.
This is not at all satisfactory.

I've seen cases where global link hints hangs, apparently because we do
not hear back from a frame.  The issue is repeatable when it happens,
but not repeatable in general (if you reload the page, then it goes
away).  This suggests that there may be something fundamentally wrong
with the global link-hints logic.  However, what?

In the interim, this adds a timer to trigger link-hints mode activation
with whataver hints we do have after a short period of time.

Ideally, we'll be able to get rid of this soon.