Do not post hint descriptors back to the frame itself.
When distributing hint descriptors, do not post a frame's own hint
descriptors back to the frame itself.  It already has them.

With regard to the message-passing cost only, this represents a speedup
of approximately 3/2 for link-busy sites like reddit -- several tens of
milliseconds for me.  There are other costs too (such as processing the
hint descriptors) bu these are not affected.