Make rects absolute in window earlier.
This is part of an effort (as yet incomplete) to track down some CSP
violations we're getting, in this case related to global link hints.

Prior to this, we were seeing CSP violations when accessing the window
subsequently to collecting the hints, specifically when accessing
window.scrollX/Y.  Here, we collect the window position immediately,
when initially harvesting the hints.