Fix Vomnibar initialization issue.
On a slow loading page, as the page is loading, hit `o` repeatedly.
Eventually, Vimium hangs.

We seem to have had this problem for quite some time (e.g. it's in
1.54).

This fixes the problem by ensuring that the Vomnibar is initialized
before launching it.

Also fix the same issue for the help dialog.