Revert always initialize help dialog.
Here's the trade off... We either *always* initialize the help dialog in
every frame of every tab, or we initialize it only when needed.

The help dialog is rarely used by anybody except beginners, so the
latter seems like the better approach.  However, on busy/slow loading
pages, the help dialog can fail to load on page start up.  (We need to
fix this, but I don't know how right now).  In the meantime, the better
compromise seems to be to initialize the help dialog only when needed.

This was the status quo prior to 6e12e005a4261711583571be23018481a4a23230
and 09e119db8d8c3ad2077ac509f73fefb76a08fac0.