Better visual-mode selection handling.
Assume there is an existing range selection.

Consider `v`, `Escape`.  Previously, this had the side effect of
clearing the selection.  That behaviour seems odd and wrong.

Here, we retain the selection if there was an existing range selection
when visual mode was launched.

Arguably, this removes a "feature" whereby `v`, `Escape` clears the
selection.  Here, we take it that the oddity above outweighs the loss of
this arguably unintentional "feature".