Simplify find-mode exit event.
Previously, we were transferring a "transferable event" from the HUD UI
component to the content script (when find mode exits).

In addition to being unnecessary, this was triggering a warning in the
console because we were reading *all* of the events keys, including one
which triggers a "this is being deprecated" warning.

The approach here is simpler, transfers less data and avoids triggering
the warning.