Cache content_scripts/vimium.css in chrome.storage.local.
Previously, we had two different approaches.  This seems like a simpler
approach.

We simply cache the Vimium CSS in chrome.storage.local (in the
background page) and fetch it from there (in UI components).

There is also a minor change in the we no longer cache the CSS in
memory.  This seems to be the right thing to do.  Caching the CSS in
memory consumes a small amount of memory.  However, it can only speed up
the fastest loads.  It cannot speed up the first load -- which is likely
the one that matters most.  So caching the CSS in memory seems to make
little sense.