Ensure help dialog is scrollable.
When we open the help dialog, scroll it, close it then open it again,
the scroller loses track of the scrollable element.  This is because the
scroller is still referenceing the old scrollable content as its
"activated element".

Here, we click the dialog element to get the scroller back in sync.

(Note:  This is a more general problem which we need to look into.)