Prime focus input on the Vimium options page.
This is a bit of a hack, and pertains to focusInput (`gf`).

The problem...  On the options page, there are *lots* of inputs
(specifically, for every exclusion rule there are two inputs).  This
makes `gf` almost unusable.  It's impossibly to `Tab` through them all.

Here, on the options page, we prime the pre-selected input to be the
key-mappings input:

- arguably, this is the most frequently used input, and
- other inputs (such as custom search engines) are just one `Tab` away.