Add sendMessage command to message external components.
This is a simpler alternative to #1980.

Example:

    map p sendMessage name=SomeName secret=SomeSecret

Now, `3p` sends `window.PostMessage` with the message:

    name: "SomeName"
    secret: "SomeSecret"
    count: 3

It is a simple matter to write a  TamperMonkey script to receive and
handle the message.  And - of course - it would be relatively simple for
users to share TamperMonkey commands.

Another example:

    map p sendMessage name=SomeName secret=SomeSecret extension=abjmlgelklomcednhcplkfdfdbajhdbn

If the `extension` option is truthy, then we use
`chrome.runtime.sendMessage()` instead, using the cross-extension
messaging API.  In this case, users could share new commands by
publishing chrome extensions.