Fix help-dialog links.
It's not clear when -- possibly when we moved the help dialog to an
iframe -- the links at the bottom of the help dialog stopped working.

Here, we reinstate them by opening the referenced page in a new tab.
This is consistent with the behaviour of the links at the top of the
help dialog.

Also, the "please review us" link now points to the "reviews" tab on the
Chrome store.