Unregister frame port on onDisconnect.
I seems that we cannot rely upon either the "unload" event in the
content script or the port's onDisconnect handler there to unregister
frames.

To see this, go to a frame-busy page like this one:

- https://www.theguardian.com/technology/2016/apr/26/apple-iphone-first-revenue-decline-13-years

and then navigate to any other simple page (in the same tab).  Navigate
back and forward with `H` and `L`.

If you watch frames registering anf unregistering, almost all of the
frames from the frame-busy page do not unregister.

Here, we unregister frames on onDisconnect in the background page too.

It is possible that this is the source of the problem mentioned as point
1 in this comment:

- https://github.com/philc/vimium/issues/2081#issuecomment-210980903

And for which #2116 is a workaround.