Match globa-mark URLs by length.
This changes the logic for selecting an existing tab when using global marks.

Previously, an exact RUL match was required.  Here, we only require a prefix match.

For example, if the global-mark URL is:

    https://inbox.google.com/u/0/

Then a tab with the URL

    https://inbox.google.com/u/0/sent

will be selected.

This is a more usable approach when the user uses global marks to visit
important sites like gmail, Facebook or Twitter.  On these sites, the
URL changes, but the user still thinks of the tab as their "gmail tab",
for example.

Also, when choosing between multiple candidate tabs:

- If there is at least one candidate in the current window, then only consider candidates in the current window.
- If there are more than one candidates, then don't select the current tab.
- Finally, select the remaining candidate with the shortest URL.

Closes #2248.