Space rotates hints (to make hidden hints visible).
It is common for link-hint markers to be close togother, and
overlapping.

Here, `<Space>` rotates hint markers such that hidden markers become
visible.

For filtered hints we additionally require a modifier (because `<space>`
on its own is already a token separator).

The calculation of overlapping hints is quite expensive: O(n^2) in the
best case and O(n^3) in the worst cast.  The worst case is
extraordinarily unlikely to arise in practice.