Disable content-script loader (temporary).
Scripts loaded via the content-script loader previously were loaded
synchronously.  In recent versions of Chrome, they now seem to be loaded
asynchrnously, which causes errors.

This is a temporary commit to keep master healthy until I (or somebody)
figures out properly what is going on.

It is expected to be reverted in due course.