Use HTML5 import to load content scripts.
This is one approach to fixing #2277 properly.  (Currently there's a
temporary fix in place.)

This uses HTML5 imports in place of the content-script loader.

This is better than the current "fix" because content scripts are only
listed in one place two places (as opposed to six).

It's worse than the previous content-script loader, though, in that
content scripts must be listed in two places, instead of one.

Fixes #2277.