Float find-mode matches to the right.
This floats the "5 Matches" text in the find-mode HUD to the right.

This looks nicer, because the "5 Matches" text doesn't move as you type.

As it happens, it also fixes an issue which arose a few months ago (for
unknown reasons) with the positioning of the cursor in find mode.  The
cursor position was becoming "jammed" after the first character as you
type.