Fix (another) infinite-scroll issue.
Steps to reproduce:

- Press and hold `j`.
- Tab `k`.
- Release `j`.

The tap on `k` is uninstalling the `cancelEventListener` installed by
`j`, and because we advance `@time`, the `j` stops scrolling and
removes the `cancelEventListener` installed for `k`.  So we end up with
no `cancelEventListener` installed.

The fix is to make the `cancelEventListener` specific to the scroller
instance.

The more fundamental problem here is that we're mixing dynamic and
static state.  A better approach would be to have `CoreScroller` as a
class, with a new instance created for each scroll instance.