Fix visual line mode line selection behavior
The current line selection behavior in the visual line mode is not
the same as in vim, where we cannot select lines upwards.

Steps to reproduce:

- Press `v` and `c` to enter the caret mode.
- Press `k`.

This is because when only the base line (the line selected when we
enter the visual line mode) is selected, and its direction is different
from the direction of the command, that command will actually unselect
the base line.

To fix this, when executing line selection commands, we check if the
base line is unselected. If so, we restore it and reverse its direction.

Test cases are added.