Move key handling to keydown and event.key.
- For PDF tabs, we see keydown events but we don't see the corresponding
  keypress events, so this allows us to navigate through PDF tabs with
  J/K.

- This relies on using `event.key` instead of `event.keyIdentifier`
  (which is no longer supported).

- Also move mapkey handling into the keyboard utils.

The effect of all of this is to simplify and unify much of our key
handling.

Fixes #1243 (for tab commands).

There are known places where we can prune a considerable amount of key
handling code.  However, that code is left in place for now in case this
change throws up some unexpected issues.