Make the consumeKeyup handler a singleton.
That is, allow at most one handler to be installed at any one time.

I have not observed this to be necessary.  However, if there were any
systematic way in which we weren't seeing the necessary keyup events,
then these handlers would just be added and added.  So this seems safer.