Fix tests for #2626.
This fixes the tests for #2626.

Note: This may not be th best approach.

The problem is that, for the first time, we're using `Settings` (and
hence `chrome.storage`) within the Vomnibar and HUD iframes, and our
`chrome` stubs are not injected into those frames.

Mention @mrmr1993.  Matt: Do you know of a better approach?  Can we
inject the stubs programmatically in the tests?  An alternative approach
would be appreciated.