Revert #2636.
Mention @mrmr1993.
Mention @gdh1995.

We have zoom issues resulting from:
   - style
   - zoom
   - --force-device-scale-factor=1.5
   - HiDPI

This intended to fix HiDPI, but seems to have created problems in the
some of the other cases.

I don't have time to figure out the problem right now, so I'm going to
revert this and push it as 1.60.2.

That leaves us pretty much where we were before, I think, in terms of
zoom.

We can come back to it later.