FF: Show backup download link in the options page
Firefox doesn't seem to respect the |download| attribute for links, and
it garbage collects our blob URL when we navigate away from the options
page (to the object URL). We work around this by simply showing the
download link in Firefox.

Disappointingly, the user needs to right-click the link and choose to
download from the context menu.

Behaviour in Chrome is unchanged, except that the download link is now
visible.