Implement a renderer
So far the renderer does not resolve overlaps (with or without z-index).
The only implemented parts are:
* detecting elements on-screen
* pruning elements that are hidden in the overflow of an ancestor
* marking elements which are not rendered but with rendered decendents

The implementation does some basic things to try and avoid extra/repeat
computations, but some things remain to be done, if required for speed.
For example, it resolves overflow for elements which may be pruned later
by the recipient of the rendering (e.g. LinkHints), which is technically
wasted computation.