FF: Restore window expando on every event listener
This should fix *most* problems from Firefox issue 1408996[1].

The root cause of the bug is that document.domain is set, and Firefox
discards the 'expando' for the window that contains all properties we
set. Technically, this can happen at any time, and so some failures are
still possible. This fixes #2799.

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=1408996